+++ apis/vara-eth/CHANGELOG.md
@@ -9,7 +9,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 
 - Expose contract addresses from mirror/router/wvara clients in https://github.com/gear-tech/gear-js/pull/2197
+- Move `InjectedTransaction` class logic into `Injected` class in https://github.com/gear-tech/gear-js/pull/2207
+- Update `IInjectedTransaction` interface with comprehensive JSDoc documentation in https://github.com/gear-tech/gear-js/pull/2207
 
 ### Added
 
 - Pick the next validator as a recipient for injected transactions in https://github.com/gear-tech/gear-js/pull/2205
+
+### Removed
+
+- `InjectedTransaction` class in https://github.com/gear-tech/gear-js/pull/2207